Refectoring test code.
